requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
cloudinary==1.44.1
dj-database-url==3.1.2
Django==5.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-jazzmin==3.0.2
django-unfold==0.80.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==25.1.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.11.0
